refactor: Remove useless 386 file

+++ stat/access_date_unix.go
@@ -1,7 +1,5 @@
-//go:build unix && !darwin && (amd64 || arm64)
-// +build unix
-// +build !darwin
-// +build amd64 arm64
+//go:build unix && !darwin
+// +build unix,!darwin
 
 package stat
 
@@ -16,6 +14,6 @@ func AccessDate(info os.FileInfo) (t time.Time, err error) {
 	if !ok {
 		return t, ErrGettingStruct
 	}
-	t = time.Unix(stat.Atim.Sec, stat.Atim.Nsec)
+	t = time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec))
 	return
 }
